Consider `deploy.environment` value when finidng default branch
This is a continuation of @39323f0, which addressed the issue of
finding the default branch name for deployment for Heroku.

In addition to the `app` key, we support the `environment` key value
in `engineyard` provider.
This was omitted in the previous commit, leading to
https://github.com/travis-ci/travis-ci/issues/6212

Thus, this commit resolves https://github.com/travis-ci/travis-ci/issues/6212